requirements.txt
selene == 2.0.0rc9
pytest == 8.3.1
allure-pytest == 2.13.5
python-dotenv == 1.0.1
requests == 2.32.3
appium-python-client == 4.0.1
